+++ src/app/(viewer)/viewer/@contents/page.tsx
@@ -1,5 +1,5 @@
+import { Unauthorized } from "@/components/card/unauthorized";
 import { Badge } from "@/components/ui/badge";
-import { Unauthorized } from "@/components/unauthorized";
 import { checkSelfAuthOrRedirectToAuth } from "@/features/auth/utils/get-session";
 import { hasContentsPermission } from "@/features/auth/utils/role";
 import { ViewerStack } from "@/features/viewer/components/viewer-stack";
