+++ src/app/error.tsx
@@ -1,5 +1,5 @@
 "use client"; // Error components must be Client Components
-import { StatusCodeView } from "@/components/status-code-view";
+import { StatusCodeView } from "@/components/card/status-code-view";
 import { Button } from "@/components/ui/button";
 import { ERROR_MESSAGES } from "@/constants";
 import { loggerError } from "@/pino";
