+++ src/app/not-found.tsx
@@ -1,4 +1,4 @@
-import { NotFound } from "@/components/not-found";
+import { NotFound } from "@/components/card/not-found";
 
 export default function Page() {
 	return <NotFound />;
