+++ src/app/search/page.tsx
@@ -1,4 +1,4 @@
-import { StatusCodeView } from "@/components/status-code-view";
+import { StatusCodeView } from "@/components/card/status-code-view";
 
 export const dynamic = "force-dynamic";
