+++ src/components/card/unauthorized.tsx
@@ -1,4 +1,4 @@
-import { StatusCodeView } from "@/components/status-code-view";
+import { StatusCodeView } from "@/components/card/status-code-view";
 import { Button } from "@/components/ui/button";
 
 export function Unauthorized() {
