+++ src/features/image/components/image-stack-provider.tsx
@@ -1,4 +1,4 @@
-import { StatusCodeView } from "@/components/status-code-view";
+import { StatusCodeView } from "@/components/card/status-code-view";
 import { ERROR_MESSAGES } from "@/constants";
 import { getUserId } from "@/features/auth/utils/get-session";
 import { generateUrlWithMetadata } from "@/features/image/actions/generate-url-with-metadata";
