+++ src/features/news/components/news-stack.tsx
@@ -1,7 +1,7 @@
 "use client";
+import { StatusCodeView } from "@/components/card/status-code-view";
 import { SmallCard } from "@/components/stack/small-card";
 import { StackSkeleton } from "@/components/stack/stack-skeleton";
-import { StatusCodeView } from "@/components/status-code-view";
 import type { News } from "@/features/news/types";
 
 type Props = {
